Step 1/6: **Project configuration and data models:** Update `ReportingDashboard.cs...
