+++ src/ReportingDashboard/ReportingDashboard.csproj
@@ -4,7 +4,6 @@
     <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
-    <RootNamespace>ReportingDashboard</RootNamespace>
   </PropertyGroup>
 
 </Project>
